Corrección menor
Se corrigió la cantidad de digitos que podía recibir la cédula jurídica a 10 ya que no se permitía más de 9 digitos

+++ database/migrations/2019_03_12_192914_create_company_table.php
@@ -15,7 +15,7 @@ class CreateCompanyTable extends Migration
     {
         Schema::create('company', function (Blueprint $table) {
             $table->string('name');
-            $table->integer('legal_id')->primary();
+            $table->unsignedBigInteger('legal_id')->primary();
             $table->text('address');
             $table->timestamps();
         });
